add supersetGroup button
wip: this feature may not actually be useful.
It currently just shows which group each record is a part of, but not
what other records are in that group. To do that each record would have to pull
all other records from the same date + group to find their exercise names.
It's also unclear whether to have the badge indicate the group vs how many other records are in the group.
Supersets don't really empirically affect any training so there may not actually be a need to include a programmatic method for tracking them.